second commit

+++ expense/src/main/java/repository/ExpenseRepository.java
@@ -3,5 +3,5 @@ package repository;
 import model.ExpenseClaim;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface ExpenseRepository extends JpaRepository<ExpenseClaim,Integer> {
+public interface ExpenseRepository extends JpaRepository<ExpenseClaim,Long> {
 }
